+++ packages/transaction-confirmation/src/index.ts
@@ -1,0 +1,5 @@
+export * from './confirmation-strategy-blockheight';
+export * from './confirmation-strategy-nonce';
+export * from './confirmation-strategy-recent-signature';
+export * from './confirmation-strategy-timeout';
+export * from './waiters';
