+++ packages/transaction-confirmation/tsconfig.json
@@ -1,0 +1,15 @@
+{
+    "$schema": "https://json.schemastore.org/tsconfig",
+    "compilerOptions": {
+        "lib": [
+            "DOM",
+            "ES2020",
+            "ES2022.Error"
+        ]
+    },
+    "display": "@solana/transaction-confirmation",
+    "extends": "tsconfig/base.json",
+    "include": [
+        "src"
+    ],
+}
